Move Monaco re-exports out of the utility barrel
Direct internal callers now import `Uri`, `KeyMod`, `KeyCode`, and
`Range` from `./utility/monaco-ssr` instead of the `./utility`
barrel. This keeps the barrel free of Monaco's heavy ESM dependency
graph, so callers that only need lightweight helpers (`cn`, `pick`,
etc.) don't transitively pull Monaco.

The package's public API is preserved: `index.ts` continues to
expose those four symbols directly.
