+++ packages/graphiql-react/src/utility/index.ts
@@ -8,7 +8,6 @@ export {
 export { debounce } from './debounce';
 export { formatJSONC, parseJSONC, tryParseJSONC } from './jsonc';
 export { markdown } from './markdown';
-export { Uri, KeyMod, KeyCode, Range } from './monaco-ssr';
 export { pick } from './pick';
 export { useDragResize } from './resize';
 export { clsx as cn } from 'clsx';
